all PAGES files updated, styles extracted, cleaned up
