+++ src/app/pages/dashboard/dashboard-page/dashboard-page-styles.tsx
@@ -1,0 +1,17 @@
+import { makeStyles } from '@material-ui/core/styles';
+export const useStyles = makeStyles((theme) => ({
+    root: {
+        display: 'flex',
+        minHeight: '100vh',
+    },
+    main: {
+        display: 'flex',
+        flexDirection: 'column',
+        flexGrow: 1,
+        backgroundColor: theme.palette.background.default,
+        padding: theme.spacing(3),
+    },
+    toolbar: {
+        [theme.breakpoints.down('sm')]: theme.mixins.toolbar,
+    },
+}));
