+++ src/app/pages/dashboard/dashboard-page/dashboard.page.tsx
@@ -1,34 +1,14 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
-
-import { makeStyles } from '@material-ui/core/styles';
-
-import { DashboardProvider } from '../../state';
-import { Heading, NavigationDrawer } from '../../components';
-
-import { HomePage } from './home/home.page';
-import { CardAddPage } from './payment/add/add.page';
-import { CardListPage } from './payment/list/list.page';
-import { SubscribePage } from './payment/subscribe/subscribe.page';
-import { EditProfilePage } from './child/edit/edit.page';
-import { CreateChildPage } from './child/create/child.page';
-
-const useStyles = makeStyles((theme) => ({
-    root: {
-        display: 'flex',
-        minHeight: '100vh',
-    },
-    main: {
-        display: 'flex',
-        flexDirection: 'column',
-        flexGrow: 1,
-        backgroundColor: theme.palette.background.default,
-        padding: theme.spacing(3),
-    },
-    toolbar: {
-        [theme.breakpoints.down('sm')]: theme.mixins.toolbar,
-    },
-}));
+import { DashboardProvider } from '../../../state';
+import { Heading, NavigationDrawer } from '../../../components';
+import { HomePage } from '../home/home.page';
+import { CardAddPage } from '../payment/add/add.page';
+import { CardListPage } from '../payment/list/list.page';
+import { SubscribePage } from '../payment/subscribe/subscribe.page';
+import { EditProfilePage } from '../child/edit/edit.page';
+import { CreateChildPage } from '../child/create/child.page';
+import { useStyles } from './dashboard-page-styles';
 
 const DashboardPage: React.FC = () => {
     const classes = useStyles();
