+++ src/app/pages/dashboard/home/home.page.tsx
@@ -1,33 +1,8 @@
 import React from 'react';
-
-import { makeStyles } from '@material-ui/core/styles';
 import { CircularProgress } from '@material-ui/core';
-
 import { ParentContext, ChildListContext } from '../../../state';
 import { ParentCard, ChildList } from '../../../components';
-
-const useStyles = makeStyles((theme) => ({
-    loading: {
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-        minHeight: '100%',
-    },
-    content: {
-        display: 'grid',
-        gridTemplate: `
-            "header header" auto
-            "children children" auto`,
-        gridTemplateColumns: '1fr 1fr',
-        gridGap: theme.spacing(3),
-    },
-    header: {
-        gridArea: 'header',
-    },
-    children: {
-        gridArea: 'children',
-    },
-}));
+import { useStyles } from './home.page.styles';
 
 interface MainPageProps {
     className?: string;
