+++ src/app/pages/dashboard/payment/list/list.page.spec.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-
 import { CardListPage } from './list.page';
 
 describe('App', () => {
