+++ src/app/pages/dashboard/payment/subscribe/subscribe.page.spec.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-
 import { ChildListProvider } from '../../../../state';
 import { SubscribePage } from './subscribe.page';
