+++ src/app/pages/dashboard/payment/subscribe/subscribe.page.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import { StripeWrapper, AddSubscription } from '../../../../components';
-
 const SubscribePage: React.FC = () => {
     const history = useHistory();
     const { id } = useParams();
