+++ src/app/pages/error-page/error-page.spec.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-
 import { ErrorPage } from './error-page';
 
 describe('ErrorPage', () => {
