+++ src/app/pages/index.ts
@@ -1,4 +1,4 @@
-export * from './dashboard/dashboard.page';
+export * from './dashboard/dashboard-page/dashboard.page';
 export * from './sign-up/sign-up.page';
 
 export * from './privacy-policy/privacy-policy.page';
@@ -14,9 +14,7 @@ export * from './pdf-display/pdf-display.page';
 
 export * from './admin-dashboard/admin-dashboard.page';
 export * from './admin-onboarding';
-
 export * from './error-page/error-page';
-
 export * from './help-page/help-page';
 export * from './child-home/child-home.page';
 export * from './versus/versus.page';
