+++ src/app/pages/pdf-display/pdf-display.page.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-
 import { PDFDisplay } from '../../components';
 
 const PdfDisplayPage: React.FC = () => {
     const { week } = useParams();
-
     if (!week) return <div></div>;
     return (
         <div>
