+++ src/app/pages/privacy-policy/privacy-policy.page.spec.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-
 import { PrivacyPage } from './privacy-policy.page';
 
 describe('PrivacyPage', () => {
