+++ src/app/pages/privacy-policy/privacy-policy.page.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-
 import { PrivacyPolicy } from '../../components';
-
 const PrivacyPage: React.FC = () => {
     return (
         <>
