+++ src/app/pages/sign-up/images/cityscape.png
