+++ src/app/pages/tos/tos.page.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import { ToS } from '../../components';
 
 const ToSPage: React.FC = () => {
