--- src/app/pages/versus/versus-notes.md
@@ -1,1 +1,0 @@
-This is the `MATCHUP` page with the route `./matchup`
