+++ src/app/pages/versus/versus-page-styles.tsx
@@ -1,5 +1,4 @@
-import { AppBar, Button, CircularProgress, Toolbar, Typography } from '@material-ui/core';
-import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 
 export const useStyles = makeStyles((theme) => ({
     loading: {
